Bugfix: Roundrobin auto-scheduler now handles gaps in allocated gameslots
Prior to this fix, an intentionally omitted week (ie: for Christmas) in the
available gameslots caused the roundrobin scheduler to break, as it only knew
how to add 7 days to the previous date and try to schedule.  Now, it uses the
next available gameslot, skipping over any gaps.

This hasn't been tested on leagues with multiple games per week, but it should
also solve that TODO item as well.