Force newly-created user's ID to 'new' temporarily.
This should fix the root cause of the issue discovered by Steve Biggs.  With a
faux userid of 'new', the edit permission checks do the right thing and allow
editing of fields.  Without it, they try to check that the current session user
has permission to edit. Since the session user does not exist when anonymous
users are creating a new account, it barfs.