Cleanup: Member ID now based on user id.
The old hack of generating a member ID based on 4 digits for the year the
player started playing, one for the gender of the player, and 3 digits for the
individual was starting to approach capacity.  OCUA shows over 700 players of
each gender starting play in 2003, and the number keeps increasing as players
who began playing in other cities move to Ottawa and create accounts.

So... instead of perpetuating this system, we will switch to a much simpler one
-- the member ID will be generated based on the user_id value plus 21000000.
The choice of constant value is to prevent collisions with existing member_id
values so that we don't have to renumber everyone.   Believe it or not, some
people have bothered to memorize their OCUA member ID.