Bugfix: Don't use rowCount on a SELECT statement.
rowCount isn't guaranteed to work on SELECT.  And, on some versions of MySQL it
clearly doesn't, breaking just about everything.  Since in most places it's
either cosmetic (in the search results) or unnecessary (in classes/*.inc) I've
removed them.

We can fix the cosmetic bits (displaying "no results" output vs an empty list
of results on the search page, for example) later.