Make it obvious that "ok" spirit is still considered good, but perfect spirit is better.
Previously, you got a checkmark icon for perfect, a lame little - for OK, and
an X for "not ok" in each questionnaire category.  Unfortunately, this seems to
have confused some OCUA players.

Now, you get a shiny gold star for perfect, a checkmark for OK, and an X for
not OK.   This should, hopefully, reduce the complaints from teams that  got a
"-" for having a OK-but-not-great game while still making it clear that they
can do better.

Icons are replaced with better looking public-domain ones from openclipart.org.
Generated .png files are provided, but the .svg source is available if
modification is necessary.