Feature: Implement automatic game-end values based on sunset time.
Previously, if an end time was neglected at gameslot creation time, games
scheduled in that slot would default to an end time of "dark".  Teams would
then be expected to keep track of a list of timecaps outside of leaguerunner.
This is unnecessary and confusing, as is the alternative of specifically
entering timecaps for games that aren't otherwise limited by field permit time
limits.

Now, if you enter your latitude, longitude, and timezone offset from GMT under
"global settings", an appropriate finish time will be calculated based on
sunset.  The time displayed is the sunset time rounded down to the nearest
5-minute interval.

TODO: if fields have a lat/long entered, use that location in preference to the
global one from settings.