Remove Google Calendar link.
The "Add to Google Calendar" link doesn't work, and none of the docs available
from Google indicate that it should.  There's no one-click addition of an
iCalendar feed to Google Calendar that I can find.  It seems that Google wants
you to individually register any external calendar into Google Calendar, and
then reexport it to other calendar users.  That's just not feasable for
Leaguerunner.