Cleanup: Don't show almost-perfect ratios in red on League Status page
Odd numbers of games cause an annoying edge case.  If, at an even number of
games, all teams are at a perfect 0.500 home/away ratio, the next game
scheduled will cause half of them to show in red as they're now below 50%.  The
same problem exists for preferred-region balancing.

So, let's not flag teams that are off perfect balance by a fractional game in
red to avoid scaring the coordinators.