Bugfix: Fix some field allocation fairness problems
Sub-fields referencing a parent probably don't have their region set, so we
need to check the parent's region value.  Otherwise, those fields are
considered to be in a null region and region matching won't pick them up.  (for
those of you following along with the home version of our game show, this is
why Ottawa's LLL 7 and 8 were rarely allocated, even though they're in Central)