Rework the internals of the SOTG system
The new system uses a single table, spirit_entry, to track per-team SOTG
entries.  This lets us have a single row per team, per game instead of the
multiple rows per submission under the old system.

Much of the code has been pulled into a 'Spirit' class that should allow for a
single place to maintain calculations and submission validation.  The previous
use of the database to generate spirit questionnaires has been replaced with
question descriptions in this class, for ease of maintenance.

There may be bugs, but this implementation should be at least as complete and
functional as what it replaces.