Implement multi-pass field assignment
Previously, field assignments were done in one pass.  If there were no
available fields in the home team's region, the away team gets their choice (if
available).  This can unfairly reward the away team in some cases -- for
example, if the home team is an 0/3 West team playing a 2/3 Central team, that
game will be chosen near the top for scheduling.  If there are enough other 0/3
West teams being scheduled before this one, the 2/3 Central away team gets
their preference, which takes a field away from the "home" Central teams
further down the list.

It all did come out in the wash later on, as the auto-balancing does trend to
50% over time, but it may take the full season to get there, and that doesn't
help coordinators when someone starts complaining mid-season.

The solution here implements multiple passes over the list of games:
        a) Assign permanent "home" fields, and remove matches from list.
        b) Assign by "home" preference, and remove matches from list.
        c) Assign by "away" preference, and remove matches from list.
        d) Assign randomly (though, see item (2) below for a better way).

Doing it this way means that "away" preferences can't starve a region out for
the home teams further down the list.  This is actually how it was attempted
several years ago, except there were so many bugs in that code, it was cleaner
to strip it all out and rely on the statistical balancing instead.