Convert all XXX_load and XXX_query methods to instead be static class methods.
Unfortunately, due to the fact that PHP is a horrible language, we can't
yet replace all of the identical "load" methods in each of the subclasses with
a single inherited static load in lrobject, as PHP's early static binding
prevents self:: from referencing the correct class.  We need the "Late static
binding" support in php 5.3.x for this, which isn't packaged by default on some
platforms.