Explicitly use a session name of "leaguerunner" for our cookie.
Problem discovered when trying to integrate with both Drupal 5 and
Drupal 7 simultaneously.  The SESS<base64> session name causes
problems because 5 uses md5 and 7 uses SHA256.  7 also resets the
cookie value after a login.

So, we avoid all that and set a second cookie for Leaguerunner in
the Drupal auth plugins.  To sanely do this, the session cookie is
now named 'leaguerunner'.